requirements.txt
django==1.11.18
djangorestframework
djangorestframework-jwt
psycopg2
coreapi